poc: test approach with collecting metrics by patching skeleton only
