+++ .circleci/config.yml
@@ -270,6 +270,8 @@ jobs:
             make docker-run-tests
       - store_test_results:
           path: target/reports/
+      - store_artifacts:
+          path: tests/integration/reports
       - persist_to_workspace:
           root:
             /tmp/workspace
